+++ src/Clerk/BackendAPI/Models/Operations/UsersGetOrganizationInvitationsRequest.cs
@@ -29,7 +29,7 @@ namespace Clerk.BackendAPI.Models.Operations
         /// </remarks>
         /// </summary>
         [SpeakeasyMetadata("queryParam:style=form,explode=true,name=limit")]
-        public double? Limit { get; set; } = 10D;
+        public long? Limit { get; set; } = 10;
 
         /// <summary>
         /// Skip the first `offset` results when paginating.<br/>
@@ -40,7 +40,7 @@ namespace Clerk.BackendAPI.Models.Operations
         /// </remarks>
         /// </summary>
         [SpeakeasyMetadata("queryParam:style=form,explode=true,name=offset")]
-        public double? Offset { get; set; } = 0D;
+        public long? Offset { get; set; } = 0;
 
         /// <summary>
         /// Filter organization invitations based on their status
